+++ src/commands/sysinfo.js
@@ -1,16 +1,34 @@
+// Get modules
 const os = require("os");
-
 const chalk = require("chalk");
 
-const _convertSize = require("../functions/convSize");
-const _fatalError = require("../functions/fatalError");
+// Get variables
 const { GLOBAL_NAME } = require("../variables/constants");
 
+// Get functions
+const _convertSize = require("../functions/convSize");
+const _fatalError = require("../functions/fatalError");
+
+/**
+ * Convert seconds to minutes, hours and days.
+ *
+ * The `recommended` value will check if the value
+ * is `0`, going down from `days` to `seconds`. It
+ * will also suffix the respective type in either
+ * plural or singular form (in an object).
+ *
+ * @param {number} seconds The seconds to convert.
+ * @param {number} decimals The number of decimal points to keep.
+ * @returns An object containing all of the conversions.
+ */
 const _convertTime = (seconds, decimals = 2) => {
+  // Convert seconds to minutes, hours and days using their respective mathematical convertion method
   const minutes = parseFloat((seconds / 60).toFixed(decimals));
   const hours = parseFloat((minutes / 60).toFixed(decimals));
   const days = parseFloat((hours / 24).toFixed(decimals));
 
+  // Calculate the recommended value by checking if the value is 0, going from days to seconds.
+  // It will also suffix the respective type in either plural or singular form (in an object).
   const recommended =
     parseFloat(days.toFixed(0)) !== 0
       ? { value: days, type: days === 1 ? "day" : "days" }
@@ -20,6 +38,7 @@ const _convertTime = (seconds, decimals = 2) => {
       ? { value: minutes, type: minutes === 1 ? "minute" : "minutes" }
       : { value: seconds, type: seconds === 1 ? "second" : "seconds" };
 
+  // Return as an object
   return {
     recommended,
     seconds,
@@ -29,7 +48,19 @@ const _convertTime = (seconds, decimals = 2) => {
   };
 };
 
+/**
+ * Determine the color depending on the percentage
+ * of memory being used.
+ *
+ * Algebra used: `(t - f) > (t / 2)`
+ *
+ * @param {string | number} mem The memory to display in the color.
+ * @returns A string with the colored string.
+ */
 const _determineColor = (mem) => {
+  // Algebra: (t - f) > (t / 2)
+  // If the total memory minus the free memory is greater than the total memory divided by two,
+  // it is using more than half of the total memory, hence, use a red color, else, use green.
   if (os.totalmem() - os.freemem() > os.totalmem() / 2) {
     return chalk.red(mem);
   } else {
@@ -37,28 +68,66 @@ const _determineColor = (mem) => {
   }
 };
 
+/**
+ * Get the friendly version of the OS from `os.type()`.
+ *
+ * @returns The friendly name.
+ */
 const _friendlyOS = () => {
-  let friendlyName = os.type();
-  if (friendlyName === "Darwin") {
-    friendlyName = "macOS";
-  } else if (friendlyName === "Windows_NT") {
-    friendlyName = "Windows";
+  switch (os.type()) {
+    // Darwin is another name for 'macOS', which many don't know
+    case "Darwin":
+      return "macOS";
+    // BubbleOS does not work on Windows 9x :)
+    case "Windows_NT":
+      return "Windows";
+    // If it does not match, return the default (in the case of Linux, for example)
+    default:
+      return os.type();
   }
-  return friendlyName;
 };
 
+/**
+ * Get system information about the computer from
+ * the BubbleOS CLI shell.
+ *
+ * Usage:
+ *
+ * ```js
+ * sysinfo(); // (Filter) Arguments available!
+ * ```
+ *
+ * Get lots of information about the local computer
+ * in this command. You can also filter it using
+ * arguments, which are listed below.
+ *
+ * Available arguments:
+ * - `-c`: Display computer information.
+ * - `-s`: Display system resources.
+ * - `-a`: Display advanced information.
+ * - `-e`: Display environment variables.
+ * - `--all`: Display all system information that is available.
+ *
+ * @param  {...string} args Arguments to modify the behaviour of `sysinfo`.
+ */
 const sysinfo = (...args) => {
   try {
-    const rmTip = args?.includes("--rm-tip") || args?.includes("/rm-tip");
-
+    // Initialize arguments
+    // Arguments to modify what is shown
     const computerInfo = args?.includes("-c") || args?.includes("/c");
-    const sysResource = args?.includes("-r") || args?.includes("/r");
+    const sysResource = args?.includes("-s") || args?.includes("/s");
     const advancedInfo = args?.includes("-a") || args?.includes("/a");
     const envVars = args?.includes("-e") || args?.includes("/e");
-    const all = args?.includes("--all") || args?.includes("/all");
 
+    // Show all values
+    const all = args?.includes("--all") || args?.includes("/all");
+    // Remove the tip
+    const rmTip = args?.includes("--rm-tip") || args?.includes("/rm-tip");
+
+    // In case no arguments were passed to modify what was shown, show that
     const defaultDisplay = !computerInfo && !sysResource && !advancedInfo && !envVars && !all;
 
+    // If the user either requested everything, just the computer info, or the default display
     if (all || computerInfo || defaultDisplay) {
       console.log(`${chalk.bold.underline("Computer Information")}`);
 
@@ -71,9 +140,11 @@ const sysinfo = (...args) => {
       console.log();
     }
 
+    // If the user either requested everything, just the system resources, or the default display
     if (all || sysResource || defaultDisplay) {
       console.log(`${chalk.bold.underline("System Resources")}`);
 
+      // Show the memory out of the total memory in the color designated
       console.log(
         `Memory usage: ${chalk.italic(
           _determineColor(
@@ -85,12 +156,14 @@ const sysinfo = (...args) => {
       );
       console.log(`CPU cores: ${chalk.italic(os.cpus().length)}`);
 
+      // Uptime in the recommended time (seconds, minutes, hours or days)
       const uptime = _convertTime(os.uptime(), 0).recommended;
       console.log(`System uptime: ${chalk.italic(`${uptime.value} ${uptime.type}`)}`);
 
       console.log();
     }
 
+    // If the user either requested everything or the advanced info
     if (all || advancedInfo) {
       console.log(`${chalk.bold.underline("Advanced Information")}`);
 
@@ -100,15 +173,13 @@ const sysinfo = (...args) => {
           os.endianness() === "BE" ? "BE (big endian)" : "LE (little endian)"
         )}`
       );
-      try {
-        console.log(
-          `Estimated default parallelism amount (program): ${chalk.italic(
-            os.availableParallelism()
-          )}`
-        );
-      } catch (err) {
-        console.log(`Estimated default parallelism amount (program): ${chalk.italic("N/A")}`);
-      }
+
+      // On some operating systems, this value will throw an error if run
+      console.log(
+        `Estimated default parallelism amount (program): ${chalk.italic(
+          typeof os.availableParallelism === "undefined" ? "N/A" : os.availableParallelism()
+        )}`
+      );
       console.log(
         `${GLOBAL_NAME} PID (process identification number): ${chalk.italic(process.pid)}`
       );
@@ -116,9 +187,11 @@ const sysinfo = (...args) => {
       console.log();
     }
 
+    // If the user either requested everything or the environment vars
     if (all || envVars) {
-      console.log(`\n${chalk.bold.underline("Environment Variables")}`);
+      console.log(`${chalk.bold.underline("Environment Variables")}`);
 
+      // Get the keys and values of all environment variables
       for (const [key, value] of Object.entries(process.env)) {
         console.log(`${chalk.green(key)}: ${value}`);
       }
@@ -126,6 +199,7 @@ const sysinfo = (...args) => {
       console.log();
     }
 
+    // If the user didn't put any filters and they didn't not request a tip
     if (defaultDisplay && !rmTip) {
       console.log(
         chalk.yellow.italic(
@@ -134,8 +208,10 @@ const sysinfo = (...args) => {
       );
     }
   } catch (err) {
+    // Unknown error
     _fatalError(err);
   }
 };
 
+// Export the function
 module.exports = sysinfo;