+++ CHANGELOG.md
@@ -39,6 +39,7 @@ _Note: This is a final beta compile build, and thus has lots of bug fixes, but n
 - Fixed an issue where the pre-boot interpreter would crash BubbleOS.
 - Fixed an issue where `mkfile` would crash if a directory passed in didn't exist.
 - Fixed an issue where characters such as `*` or `+` would crash BubbleOS if entered.
+- Renamed the `-r` flag in `sysinfo` to `-s` (system resources filter flag).
 - Fixed an issue where the `bub` command would report plain text files as an '_invalid encoding_'.
 - Internally renamed `aboutConsts.js` to `constants.js`.